# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation for my application",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/car/{id}": {
            "get": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Get a single car",
                "operationId": "getCarById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Update a car",
                "operationId": "updateCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand",
                                    "model",
                                    "year",
                                    "price",
                                    "manufacture"
                                ],
                                "properties": {
                                    "brand": {
                                        "type": "string",
                                        "example": "Hundai"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Tucson"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": "2024"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "45000"
                                    },
                                    "manufacture": {
                                        "type": "integer",
                                        "example": "2020"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Record not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Delete a car",
                "operationId": "deleteCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "title": "Car",
                "required": [
                    "id",
                    "brand",
                    "model",
                    "year",
                    "price",
                    "manufacturing"
                ],
                "properties": {
                    "id": {
                        "description": "Car ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "brand": {
                        "description": "Car's brand",
                        "type": "string"
                    },
                    "model": {
                        "description": "Car's model",
                        "type": "string"
                    },
                    "year": {
                        "description": "Car's year",
                        "type": "integer"
                    },
                    "price": {
                        "description": "Car's price",
                        "type": "integer"
                    },
                    "manufacturing": {
                        "description": "Car's manufacturing",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}